pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geoimporter"
version = "0.1.0"
description = "A GUI application for importing geospatial data to GeoServer"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "attrs>=23.2.0",
    "branca>=0.7.2",
    "certifi>=2024.6.2",
    "charset-normalizer>=3.3.2",
    "click>=8.1.7",
    "click-plugins>=1.1.1",
    "cligj>=0.7.2",
    "contourpy>=1.2.1",
    "cycler>=0.12.1",
    "fiona>=1.9.6",
    "folium>=0.16.0",
    "fonttools>=4.53.0",
    "geoalchemy2>=0.18.0",
    "geopandas>=0.14.4",
    "geoserver-rest>=2.7.0",
    "greenlet>=3.0.3",
    "idna>=3.7",
    "jinja2>=3.1.4",
    "joblib>=1.4.2",
    "kiwisolver>=1.4.5",
    "mapclassify>=2.6.1",
    "markupsafe>=2.1.5",
    "matplotlib>=3.8.4",
    "munkres>=1.1.4",
    "networkx>=3.3",
    "numpy>=1.26.4",
    "packaging>=24.1",
    "pandas>=2.2.2",
    "patsy>=0.5.6",
    "pillow>=10.3.0",
    "psycopg2>=2.9.9",
    "pygments>=2.18.0",
    "pyparsing>=3.1.2",
    "pyproj>=3.6.1",
    "pysocks>=1.7.1",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.1",
    "pytz>=2024.1",
    "requests>=2.32.3",
    "rtree>=1.2.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
    "shapely>=2.0.4",
    "six>=1.16.0",
    "sqlalchemy>=2.0.30",
    "statsmodels>=0.14.2",
    "threadpoolctl>=3.5.0",
    "tk>=0.1.0",
    "typing-extensions>=4.12.2",
    "urllib3>=2.2.1",
    "xmltodict>=0.13.0",
    "xyzservices>=2024.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
geoimporter = "geoimporter.geoimporter:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["geoimporter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
